Delete by email

+++ src/optional-functional/optional-functional.service.ts
@@ -19,21 +19,21 @@ export class OptionalFunctionalService {
     private readonly complianceInfoRepository: Repository<DigifranchiseComplianceInfo>,
   ) {}
 
-  async deleteUserByPhoneNumber(phoneNumber: string): Promise<void> {
-    if (!phoneNumber) {
+  async deleteUserByConnectNumber(connectNumber: string): Promise<void> {
+    if (!connectNumber) {
       throw new Error("Connect number must be provided");
     }
 
     const generalInfo = await this.generalInfoRepository.findOne({
       where: [
-        { connectNumberWithOutCountryCode: phoneNumber },
-        { otherMobileNumberWithOutCountryCode: phoneNumber },
+        { connectNumberWithOutCountryCode: connectNumber },
+        { otherMobileNumberWithOutCountryCode: connectNumber },
       ],
     });
 
     if (!generalInfo) {
       throw new NotFoundException(
-        `DigifranchiseGeneralInfo with connect number or other mobile number ${phoneNumber} not found`,
+        `DigifranchiseGeneralInfo with connect number or other mobile number ${connectNumber} not found`,
       );
     }
 
@@ -56,6 +56,48 @@ export class OptionalFunctionalService {
     }
 
     await this.franchiseOwnerRepository.remove(franchiseOwner);
+    await this.userRepository.remove(user);
+  }
+
+  async deleteByUserEmail(email: string): Promise<void> {
+    if (!email) {
+      throw new Error("Email must be provided");
+    }
+
+    const user = await this.userRepository.findOne({ where: { email } });
+    if (!user) {
+      throw new NotFoundException(`User with email ${email} not found`);
+    }
+
+    const franchiseOwner = await this.franchiseOwnerRepository.findOne({
+      where: {
+        userId: Equal(user.id),
+        userEmail: email,
+      },
+    });
+
+    if (!franchiseOwner) {
+      throw new NotFoundException(
+        `DigifranchiseOwner with email ${email} not found`,
+      );
+    }
+
+    await this.franchiseOwnerRepository.remove(franchiseOwner);
+    await this.userRepository.remove(user);
+  }
+
+  async deleteByUserPhoneNumber(phoneNumber: string): Promise<void> {
+    if (!phoneNumber) {
+      throw new Error("Phone Number must be provided");
+    }
+
+    const user = await this.userRepository.findOne({ where: { phoneNumber } });
+    if (!user) {
+      throw new NotFoundException(
+        `User with Phone Number ${phoneNumber} not found`,
+      );
+    }
+
     await this.userRepository.remove(user);
   }
 }
